backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
google-generativeai
python-multipart
python-dotenv
aiofiles
